Further menu improvements
- Fixed pack switch still getting stuck in some circumstances I cannot figure out (assignement is now constantly updated instead of being bound to key press/release events);

- Added wrapping to the level description. It would be better if pack makers updated their jsons with newlines to fit the space available as they see fit, but since it is possible some packs may never be updated I enforce a formatting when the level is changed.
It should be noted that I have noticed some weird glitches in the formatted text rendering in certain packs (particularly Babagon). Best I can tell it's due to the editor used to make the jsons since if I retype the description as in the original in my text editors the issue is fixed.